Move CreateVoyagerSchema() inside the TargetYugabyteDB.
